+++ elastalert/loaders.py
@@ -15,6 +15,7 @@ from jinja2 import FileSystemLoader
 from . import alerts
 from . import enhancements
 from . import ruletypes
+from .alerters.email import EmailAlerter
 from .alerters.opsgenie import OpsGenieAlerter
 from .alerters.zabbix import ZabbixAlerter
 from .util import dt_to_ts
@@ -59,7 +60,7 @@ class RulesLoader(object):
 
     # Used to map names of alerts to their classes
     alerts_mapping = {
-        'email': alerts.EmailAlerter,
+        'email': EmailAlerter,
         'jira': alerts.JiraAlerter,
         'opsgenie': OpsGenieAlerter,
         'stomp': alerts.StompAlerter,
